adding util.py file, renaming remaining modules
